feat(login): Bug fixes
