+++ index.js
@@ -9,22 +9,23 @@ app.commandLine.appendSwitch('auto-detect', 'false');
 app.commandLine.appendSwitch('no-proxy-server');
 
 app.on('ready', () => {
-    const mainWindow = new electron.BrowserWindow({
-        width: 1150,
-        height: 630,
-        minWidth: 950,
-        minHeight: 500,
-        resizable: true,
-        frame: false,
-        show: false,
-        title: "Lazap",
-        webPreferences: {
-            nodeIntegration: true,
-            contextIsolation: false,
-            backgroundThrottling: false,
-        },
-        icon: "icon.ico"
-    });
+	const mainWindow = new electron.BrowserWindow({
+		width: 1150,
+		height: 630,
+		minWidth: 950,
+		minHeight: 500,
+		resizable: true,
+		frame: false,
+		show: false,
+		title: "Lazap",
+		opacity: fs.existsSync(__dirname + '\\storage\\launcherprofile.json') ? require('./storage/launcherprofile.json')?.opacity : .89,
+		webPreferences: {
+			nodeIntegration: true,
+			contextIsolation: false,
+			backgroundThrottling: false,
+		},
+		icon: "icon.ico"
+	});
 	mainWindow.loadFile('src/login.html');
 
 	mainWindow.once('ready-to-show', () => {
@@ -32,15 +33,24 @@ app.on('ready', () => {
 		setTimeout(() => {
 			mainWindow.show();
 		}, 100);
-		mainWindow.webContents.send('check-if-logged-in');
 	});
 
-	mainWindow.webContents.on('did-finish-load', () => {
+	mainWindow.webContents.on('did-finish-load', async () => {
+		// Checking for auto-login
+		if (fs.existsSync(__dirname + '\\storage\\userprofile.js')) {
+			if (require('./storage/userprofile.json').token) {
+				mainWindow.webContents.send('check-if-logged-in', await identify());
+			}
+			else {
+				return;
+			}
+		};
 		handleStorageAndTransportData(mainWindow);
 	});
 
-	ipcMain.on('load-main', () => {
+	ipcMain.on('load-main', (e, data) => {
 		mainWindow.loadFile('src/index.html');
+		setTimeout(async () => mainWindow.webContents.send('check-for-login', data || await identify()), 2000);
 	});
 
 	ipcMain.on('load-custom', (e, str) => {
@@ -123,7 +133,7 @@ async function handleSignup(data) {
 async function handleSignin(data) {
 	let deniedCode;
 	const res = await axios.post('http://localhost:3000/accounts/login', data).catch(e => {
-		deniedCode = e.response?.status || 0;
+		deniedCode = e.response?.status;
 	});
 
 	if (res) {
@@ -131,10 +141,12 @@ async function handleSignin(data) {
 		if (fs.existsSync(__dirname + '\\storage\\userprofile.json')) {
 			const json = require(`${__dirname}/storage/userprofile.json`);
 			json['token'] = res.data;
+			json['password'] = data.pass;
+			json['username'] = data.username;
 			editLocalStorage(json);
 		} else {
 			editLocalStorage({
-				token: res.data
+				token: res.data,
 			});
 		}
 
@@ -142,4 +154,21 @@ async function handleSignin(data) {
 	} else {
 		return deniedCode || null;
 	}
+}
+
+async function identify() {
+	let res;
+
+	if (!fs.existsSync(__dirname + '\\storage\\userprofile.json')) return { status: 'ACCOUNT_NOT_FOUND', data: null };
+	const { token, password, username } = JSON.parse(fs.readFileSync(__dirname + '\\storage\\userprofile.json').toString());
+	res = await axios.post('http://localhost:3000/accounts/identify', { token, pass: password, name: username }).catch((e) => e.response?.status || 0);
+
+	const errcodes = {
+		401: 'INVALID_BODY',
+		402: 'ACCOUNT_NOT_FOUND',
+		200: 'SUCCESS',
+		0: 'OFFLINE/API_DOWN'
+	};
+
+	return { data: res.data, status: errcodes[res.request?.res.statusCode] || 'OFFLINE/API_DOWN' };
 }
