+++ src/login.html
@@ -26,7 +26,7 @@
                     <input type="text" placeholder="Username" />
                     <input type="email" placeholder="Email" />
                     <input type="password" placeholder="Password" />
-                    <button onclick="load_main()">Sign Up</button>
+                    <button>Sign Up</button>
                 </div>
             </div>
             <div class="form-container sign-in-container">
@@ -34,7 +34,7 @@
                     <h1>Sign in</h1>
                     <input type="text" placeholder="Username" />
                     <input type="password" placeholder="Password" />
-                    <button onclick="load_main()">Sign In</button>
+                    <button>Sign In</button>
                     <a>Forgot your password?</a>
                     
                 </div>
